Add offline persistence implementation, refactor to offline first, minor ui fixes

+++ app/src/main/java/com/example/resttestapp/data/models/LocalListItemModel.kt
@@ -13,14 +13,14 @@ data class LocalListItemModel(
     var userId: Number,
 ) {
     fun toLocalListItemEntity(): LocalListItemEntity = LocalListItemEntity(
-        id = id,
-        photoAlbumId = photoAlbumId,
+        id = id.toInt(),
+        photoAlbumId = photoAlbumId.toInt(),
         photoThumbnailUrl = photoThumbnailUrl,
         photoTitle = photoTitle,
         photoUrl = photoUrl,
         postBody = postBody,
         postTitle = postTitle,
-        userId = userId
+        userId = userId.toInt()
     )
 }
